refactor: create layout for events
